Refine racadmsim and template
1. racadmsim package update

`api.py`
Provide interactive shell operation, possibly transferred to some
API invoke.

`model.py`
Responsible for get updated with unified model.

`env.py`
Environment variable for this racadmsim instance.

2. racadm jinja template

Added template for storage and storage information in hwinventory.

Added `macros.j2` as macro library.
