+++ infrasim/racadmsim/env.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+'''
+*********************************************************
+Copyright @ 2015 EMC Corporation All Rights Reserved
+*********************************************************
+'''
+
+auth_map = {}
+racadm_data = None
+r_log = None
+node_name = None
