+++ template/racadmsim/hwinventory_storage_drive_satadom.j2
@@ -1,0 +1,36 @@
+
+[InstanceID: Disk.Direct.5-3:AHCI.Embedded.2-1]
+Device Type = PhysicalDisk
+LastUpdateTime = 2016-07-08T01:31:14
+LastSystemInventoryTime = 2016-12-12T09:30:09
+RemainingRatedWriteEndurance = Unknown
+OperationPercentComplete = 0%
+OperationName = None
+DriveFormFactor = 2.5 inch
+SASAddress = Not Applicable
+axCapableSpeed = 6Gbs
+UsedSizeInBytes = 0 Bytes
+FreeSizeInBytes = 0 Bytes
+ediaType = Solid State Drive
+BlockSizeInBytes = 512 Bytes
+T10PICapability = Not supported
+SecurityState = Unknown
+PredictiveFailureState = Smart Alert Absent
+HotSpareStatus = No
+BusProtocol = SATA
+SerialNumber = {{ satadom.serial }}
+Revision = {{ satadom.version }}
+anufacturingYear = 1962
+anufacturingWeek = 32
+anufacturingDay = 255
+odel = SATADOM ML 3SE
+SizeInBytes = 64023257088 Bytes
+Slot = 3
+Connector = 5
+RaidStatus = Non-RAID
+RollupStatus = Unknown
+PrimaryStatus = Unknown
+DeviceDescription = Disk 3 on Embedded AHCI Controller 2
+FQDD = Disk.Direct.5-3:AHCI.Embedded.2-1
+InstanceID = Disk.Direct.5-3:AHCI.Embedded.2-1
+-------------------------------------------------------------------
