+++ template/racadmsim/hwinventory_storage_drive_scsi.j2
@@ -1,0 +1,40 @@
+{% import "macros.j2" as lib %}
+
+
+[InstanceID: Disk.Bay.{{ drive.slot }}:Enclosure.Internal.0-1:NonRAID.Integrated.1-{{ controller_idx }}]
+Device Type = PhysicalDisk
+LastUpdateTime = 2016-12-20T08:37:44
+LastSystemInventoryTime = 2016-12-20T08:37:44
+RemainingRatedWriteEndurance = 0%
+OperationPercentComplete = 0%
+OperationName = None
+DriveFormFactor = 3.5 inch
+PPID = {{ lib.get_part_number(drive) }}
+SASAddress = {{ lib.get_wwn(drive, prefix=false) }}
+axCapableSpeed = 12Gbs
+UsedSizeInBytes = 0 Bytes
+FreeSizeInBytes = 0 Bytes
+ediaType = {{ lib.get_media_type(drive) }}
+BlockSizeInBytes = 512 Bytes
+T10PICapability = Not supported
+SecurityState = Unknown
+PredictiveFailureState = Smart Alert Absent
+HotSpareStatus = No
+BusProtocol = SAS
+SerialNumber = {{ drive.serial }}
+Revision = {{ drive.version }}
+anufacturingYear = 2016
+anufacturingWeek = 9
+anufacturingDay = 2
+anufacturer = {{ drive.vendor }}
+odel = {{ drive.product }}
+SizeInBytes = 4000787029504 Bytes
+Slot = {{ drive.slot }}
+Connector = 0
+RaidStatus = Non-RAID
+RollupStatus = OK
+PrimaryStatus = OK
+DeviceDescription = Disk {{ drive.slot }} in Backplane 1 of Integrated Storage Controller {{ controller_idx }}
+FQDD = Disk.Bay.{{ drive.slot }}:Enclosure.Internal.0-1:NonRAID.Integrated.1-{{ controller_idx }}
+InstanceID = Disk.Bay.{{ drive.slot }}:Enclosure.Internal.0-1:NonRAID.Integrated.1-{{ controller_idx }}
+-------------------------------------------------------------------
