+++ template/racadmsim/storage.j2
@@ -1,4 +1,6 @@
 {% include "storage_drive_satadom.j2" ignore missing with context %}
-{% for drive in drives %}
+{% for controller_idx, drives in mapping.iteritems() %}
+{% for drive in drives%}
 {% include "storage_drive_scsi.j2" ignore missing with context %}
+{% endfor %}
 {% endfor %}
