+++ template/racadmsim/storage_drive_satadom.j2
